app/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django==5.0.6",
    "django-environ==0.12.0",
    "djangorestframework==3.15.1",
    "psycopg[binary]==3.2.10",
]
